Update main.py
Changed the code to flask api so you can send requests to it with url and sitekey this makes it easier to integrate with other codes. I also changed the system now when flask api starts it starts the browser and waits which makes the solution faster and reduces the waiting time.